user and token auth added to delete request

+++ controllers/blogs.js
@@ -12,22 +12,12 @@ blogsRouter.get('/', async (request, response) => {
     response.json(blogs.map(blog => blog.toJSON()))
 })
 
-const getTokenFrom = request => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.toLowerCase().startsWith('bearer')){
-        return authorization.substring(7)
-    }
-    return null
-}
 
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
-    const token = getTokenFrom(request)
-    const decodedToken = jwt.verify(token, process.env.SECRET)
-    if (!decodedToken.id) {
-        return response.status(401).json({ error: 'token missing or invalid' })
-    }
 
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
     const user = await User .findById(decodedToken.id)
 
     if (!request.body.url || !request.body.title) {
@@ -51,9 +41,24 @@ blogsRouter.post('/', async (request, response) => {
     response.json(savedBlog.toJSON())
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+blogsRouter.delete('/:id', async (request, response, next) => {
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+
+    const user = await User.findById(decodedToken.id)
+
+    const blog = await Blog.findById(request.params.id)
+
+    if ( blog.user._id.toString() === user._id.toString() ) {
+        try {
+            await Blog.findByIdAndRemove(request.params.id)
+            response.status(204).end()
+        } catch (exception) {
+            next(exception)
+        }
+    } else {
+        return response.status(401).json({ error: 'Unauthorized' })
+    }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
